django_library/src/requirements.txt
Django==3.0.4
pymysql==0.9.3
mysqlclient==1.4.6
gunicorn==20.0.4
requests==2.23.0
Pillow==7.1.1
PyJWT==1.7.1
django-cors-headers==3.2.1
django-crispy-forms==1.9
djangorestframework==3.11.0
itsdangerous==1.1.0
flower==0.9.4
redis==3.4.1
celery==4.4.2
django-redis==4.11.0
djangorestframework-jwt==1.11.0
drf-extensions==0.3.1
